+++ src/main/java/org/thoughtcrime/securesms/preferences/ChatsPreferenceFragment.java
@@ -81,11 +81,7 @@ public class ChatsPreferenceFragment extends ListSummaryPreferenceFragment {
       autoDelServer.setOnPreferenceChangeListener(new AutodelChangeListener("delete_server_after"));
     }
     if (dcContext.isChatmail()) {
-      autoDelServer.setEntries(new CharSequence[]{getString(R.string.automatic), getString(R.string.autodel_at_once)});
-      autoDelServer.setEntryValues(new CharSequence[]{"0", "1"});
-      if (dcContext.getConfigInt("delete_server_after") > 1) {
-        dcContext.setConfigInt("delete_server_after", 0 /*never/automatic*/);
-      }
+      autoDelServer.setVisible(false);
     }
   }
 
@@ -277,7 +273,7 @@ public class ChatsPreferenceFragment extends ListSummaryPreferenceFragment {
                       .setTitle(R.string.pref_backup)
                       .setMessage(R.string.pref_backup_export_explain)
                       .setNeutralButton(android.R.string.cancel, null)
-                      .setPositiveButton(requireActivity().getString(R.string.pref_backup_export_x, addr), (dialogInterface, i) -> startImexOne(DcContext.DC_IMEX_EXPORT_BACKUP));
+                      .setPositiveButton(requireActivity().getString(R.string.pref_backup_export_this, addr), (dialogInterface, i) -> startImexOne(DcContext.DC_IMEX_EXPORT_BACKUP));
               int[] allAccounts = DcHelper.getAccounts(requireActivity()).getAll();
               if (allAccounts.length > 1) {
                 String exportAllString = requireActivity().getString(R.string.pref_backup_export_all, allAccounts.length);