+++ src/main/java/org/thoughtcrime/securesms/qr/QrCodeHandler.java
@@ -55,6 +55,7 @@ public class QrCodeHandler {
         switch (qrParsed.getState()) {
             case DcContext.DC_QR_ASK_VERIFYCONTACT:
             case DcContext.DC_QR_ASK_VERIFYGROUP:
+            case DcContext.DC_QR_ASK_JOIN_BROADCAST:
                 showVerifyContactOrGroup(activity, builder, rawString, qrParsed, name, source, uiPath);
                 break;
 
@@ -121,19 +122,12 @@ public class QrCodeHandler {
                 builder.setCancelable(false);
                 break;
 
-            case DcContext.DC_QR_WEBRTC:
-                builder.setMessage(activity.getString(R.string.videochat_instance_from_qr, qrParsed.getText1()));
-                builder.setPositiveButton(R.string.ok, (dialog, which) -> {
-                    dcContext.setConfigFromQr(rawString);
-                });
-                builder.setNegativeButton(R.string.cancel, null);
-                builder.setCancelable(false);
-                break;
-
             case DcContext.DC_QR_WITHDRAW_VERIFYCONTACT:
             case DcContext.DC_QR_WITHDRAW_VERIFYGROUP:
+            case DcContext.DC_QR_WITHDRAW_JOINBROADCAST:
                 String message = qrParsed.getState() == DcContext.DC_QR_WITHDRAW_VERIFYCONTACT ? activity.getString(R.string.withdraw_verifycontact_explain)
-                                  : activity.getString(R.string.withdraw_verifygroup_explain, qrParsed.getText1());
+                               : qrParsed.getState() == DcContext.DC_QR_WITHDRAW_VERIFYCONTACT ? activity.getString(R.string.withdraw_verifygroup_explain, qrParsed.getText1())
+                               : activity.getString(R.string.withdraw_joinbroadcast_explain, qrParsed.getText1());
                 builder.setTitle(R.string.qrshow_title);
                 builder.setMessage(message);
                 builder.setNeutralButton(R.string.reset, (dialog, which) -> {
@@ -146,6 +140,7 @@ public class QrCodeHandler {
 
             case DcContext.DC_QR_REVIVE_VERIFYCONTACT:
             case DcContext.DC_QR_REVIVE_VERIFYGROUP:
+            case DcContext.DC_QR_REVIVE_JOINBROADCAST:
                 builder.setTitle(R.string.qrshow_title);
                 builder.setMessage(activity.getString(R.string.revive_verifycontact_explain));
                 builder.setNeutralButton(R.string.revive_qr_code, (dialog, which) -> {
@@ -242,6 +237,9 @@ public class QrCodeHandler {
         switch (qrParsed.getState()) {
             case DcContext.DC_QR_ASK_VERIFYGROUP:
                 msg = activity.getString(R.string.qrscan_ask_join_group, qrParsed.getText1());
+                break;
+            case DcContext.DC_QR_ASK_JOIN_BROADCAST:
+                msg = activity.getString(R.string.qrscan_ask_join_channel, qrParsed.getText1());
                 break;
             default:
                 msg = activity.getString(R.string.ask_start_chat_with, name);