+++ CHANGELOG.md
@@ -1,5 +1,42 @@
 # Delta Chat Android Changelog
 
+## Unreleased
+
+* Change color of links in text messages
+* Allow to withdraw channel invite links and QR codes
+* Do not show "1 member" when the process of joining the group is not finished
+* Improve handling of video recoding
+* Send .webm videos as file, they are not supported by all platforms
+* Tweak advanced section and wording of some advanced options
+* Fix: avoid crash in push notifications handling
+* Expose new "Multi-device mode" option instead of "Delete from server" for chatmail profiles
+* Opened in-chat apps got a 'About Apps' menu item
+* Avoid gray avatar on profile creation
+* Update to core 2.27.0
+
+## v2.25.0
+2025-11
+
+* Make it possible to invite members into a channel via a QR code,
+  and make channels more secure
+* metadata protection: protect Date header
+* metadata protection: protect Autocrypt header
+* better multi-device: synchronize group creation across devices
+* data saving: do not send Autocrypt header in read receipts
+* improve onboarding speed
+* allow to save to storage files shared from inside in-chat apps
+* reduce app size
+* don't show badge counter in app icon for the permanent background notification
+* fix sorting of old media in gallery
+* fix text direction in "x members" subtitle for RTL languages
+* fix group invite QR screen's layout
+* tweak text hints in advanced classic e-mail configuration
+* remove deprecated "companion app" code
+* remove deprecated "Watch Sent Folder" preference
+* remove deprecated "send self-report" preference
+* don't show email address in shared vcard
+* update to core 2.25.0
+
 ## v2.22.0
 2025-10